Allow image cleanup before packaging
In the offline gold image builder, add config to enable DISM
Cleanup-Image to be run before packaging the image, so that
updates installed during the extra_packages phase can be cleaned
up before packaging.

The following new config options are added:
  * updates:clean_updates_offline, default false. If set,
    dism cleanup-image will be performed offline, after
    any offline updates / components / capabilities were installed.
  * updates:clean_updates_online, default true. If set,
    dism cleanup-image will be performed while the instance is running,
    after the updates were installed.

+++ UnattendResources/Logon.ps1
@@ -445,6 +445,10 @@ try {
         $enableAlwaysActiveMode = Get-IniFileValue -Path $configIniPath -Section "DEFAULT" -Key "enable_active_mode" `
             -Default $false -AsBoolean
     } catch{}
+    try {
+        $cleanUpdatesOnline = Get-IniFileValue -Path $configIniPath -Section "updates" -Key "clean_updates_online" `
+            -Default $true -AsBoolean
+    } catch{}
 
     if ($productKey) {
         License-Windows $productKey
@@ -455,14 +459,16 @@ try {
         Install-WindowsUpdates
     }
 
-    try {
-        ExecRetry {
-            Clean-WindowsUpdates -PurgeUpdates $purgeUpdates
+    if ($cleanUpdatesOnline) {
+        try {
+            ExecRetry {
+                Clean-WindowsUpdates -PurgeUpdates $purgeUpdates
+            }
+        } catch {
+            Write-Log "DISM" "Failed to cleanup updates. Rebooting..."
+            Restart-Computer -Force
+            exit 0
         }
-    } catch {
-        Write-Log "DISM" "Failed to cleanup updates. Rebooting..."
-        Restart-Computer -Force
-        exit 0
     }
 
     Run-CustomScript "RunAfterWindowsUpdates.ps1"
